requirements.txt
aio-timers==0.0.1
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
appdirs==1.4.3
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.4.0
asyncprawcore==2.3.0
attrs==20.3.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
dateparser==1.0.0
discord.py==1.7.1
distlib==0.3.0
distro==1.4.0
emoji==1.2.0
geographiclib==1.50
geopy==2.1.0
google==3.0.0
google-auth==1.29.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
html5lib==1.0.1
httplib2==0.19.1
idna==2.10
ipaddr==2.2.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.4.2
msgpack==0.6.2
multidict==5.1.0
numpy==1.20.2
oauthlib==3.1.0
packaging==20.3
pandas==1.2.4
pep517==0.8.2
Pillow==8.3.1
prawcore==2.3.0
progress==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
pytoml==0.1.21
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
table2ascii==0.0.3
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.4
webencodings==0.5.1
websocket-client==1.2.0
yarl==1.6.3
